Moved some views from the 'Start' folder to the 'CreateWallet' folder, to better represent their usage
