+++ Discreet-GUI/WPF/Views/CreateWallet/WalletCreatedSuccessfullyView.axaml
@@ -3,7 +3,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
-             x:Class="WPF.Views.Start.WalletCreatedSuccessfullyView"
+             x:Class="WPF.Views.CreateWallet.WalletCreatedSuccessfullyView"
 			 Design.Height="800" Design.Width="800">
 
   <UserControl.Styles>
