+++ Discreet-GUI/WPF/Views/CreateWallet/YourRecoveryPhraseView.axaml
@@ -2,9 +2,9 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 xmlns:vm="using:WPF.ViewModels.Start"
+			 xmlns:vm="using:WPF.ViewModels.CreateWallet"
              mc:Ignorable="d"
-             x:Class="WPF.Views.Start.YourRecoveryPhraseView">
+             x:Class="WPF.Views.CreateWallet.YourRecoveryPhraseView">
 
 	
   <UserControl.Styles>
